+++ .scripts/adb-get-db.sh
@@ -1,3 +1,11 @@
 #!/bin/sh
-# Retrieve the `tracks-info.sqlite3` database from an Android VM to a local file
-adb shell "run-as team.march.march_tales_app cat databases/tracks-info.sqlite3" > databases/db-tracks-info.sqlite3
+# Retrieve a `tracks-info.sqlite3` database from the Android VM to a local file
+mkdir -p .databases
+APP_NAME="team.march.march_tales_app"
+ADB_DB_NAME="databases/tracks-info.sqlite3"
+DATE_ID=`date "+%y%m%d-%H%M"`
+LOCAL_DB_FOLDER=".databases"
+LOCAL_DB_NAME="$LOCAL_DB_FOLDER/db-tracks-info-$DATE_ID.sqlite3"
+echo "Importing the database $ADB_DB_NAME..."
+adb shell "run-as $APP_NAME cat $ADB_DB_NAME" > "$LOCAL_DB_NAME"
+ls -lah "$LOCAL_DB_FOLDER" | grep "$DATE_ID"