+++ lib/features/Track/db/TrackInfo.dart
@@ -11,28 +11,34 @@ final formatter = YamlFormatter();
 class TrackInfo {
   int id; // track.id
   bool favorite;
-  int playedCount; // track.played_count (but only for current user!).
+  int totalPlayedCount; // track.played_count
+  int localPlayedCount; // userTrack.played_count (only for current user!).
   Duration position; // position?.inMilliseconds ?? 0
   DateTime lastUpdated; // DateTime.now().millisecondsSinceEpoch <-> DateTime.fromMillisecondsSinceEpoch(ms)
   DateTime lastPlayed; // DateTime.now().millisecondsSinceEpoch <-> DateTime.fromMillisecondsSinceEpoch(ms)
+  DateTime lastFavorited; // DateTime.now().millisecondsSinceEpoch <-> DateTime.fromMillisecondsSinceEpoch(ms)
 
   TrackInfo({
     required this.id,
     required this.favorite,
-    required this.playedCount,
+    required this.totalPlayedCount,
+    required this.localPlayedCount,
     required this.position,
     required this.lastUpdated,
     required this.lastPlayed,
+    required this.lastFavorited,
   });
 
   Map<String, dynamic> toMap() {
     return {
       'id': this.id,
       'favorite': this.favorite ? 1 : 0,
-      'playedCount': this.playedCount,
+      'totalPlayedCount': this.totalPlayedCount,
+      'localPlayedCount': this.localPlayedCount,
       'position': this.position.inMilliseconds / 1000,
       'lastUpdatedMs': this.lastUpdated.millisecondsSinceEpoch,
       'lastPlayedMs': this.lastPlayed.millisecondsSinceEpoch,
+      'lastFavoritedMs': this.lastFavorited.millisecondsSinceEpoch,
     };
   }
 
@@ -42,11 +48,16 @@ class TrackInfo {
       final trackInfo = TrackInfo(
         id: data['id'],
         favorite: data['favorite'] == 0 ? false : true,
-        playedCount: data['playedCount'],
+        totalPlayedCount: data['totalPlayedCount'],
+        localPlayedCount: data['localPlayedCount'],
         position: Duration(milliseconds: positionMs),
         lastUpdated: DateTime.fromMillisecondsSinceEpoch(data['lastUpdatedMs']),
         lastPlayed: DateTime.fromMillisecondsSinceEpoch(data['lastPlayedMs']),
+        lastFavorited: DateTime.fromMillisecondsSinceEpoch(data['lastFavoritedMs']),
       );
+      logger.t(
+          '[TrackInfo:fromMap] lastPlayedMs=${data['lastPlayedMs']} lastPlayed=${trackInfo.lastPlayed} data=${data} trackInfo=${trackInfo}');
+      // debugger();
       return trackInfo;
     } catch (err, stacktrace) {
       final String msg = 'Can not parse track info data: $err';
@@ -59,6 +70,6 @@ class TrackInfo {
 
   @override
   String toString() {
-    return 'TrackInfo(id=$id favorite=${favorite} position=${position} playedCount=${playedCount} lastUpdated=${lastUpdated} lastPlayed=${lastPlayed})';
+    return 'TrackInfo(id=$id favorite=${favorite} position=${position} totalPlayedCount=${totalPlayedCount} localPlayedCount=${localPlayedCount} lastUpdated=${lastUpdated} lastPlayed=${lastPlayed} lastFavorited=${lastFavorited})';
   }
 }