+++ CHANGELOG.md
@@ -1,9 +1,30 @@
 <!--
  @since 2025.03.19, 22:43
- @changed 2025.03.19, 22:43
+ @changed 2025.04.15, 15:53
 -->
 
 # CHANGELOG
+
+## [v.0.1.10](https://github.com/lilliputten/march-tales-flutter-app/releases/tag/v.0.1.10) - 2025.04.15
+
+- Added support for a UserTrack data inside the Track server objects.
+- Added `lastFavoritedMs` field to a local database.
+- Added invocations of `updateLocalTracks` or `updateLocalTrack` when tracks' data received.
+- Added synchronization of local track' states (positions, favorite) on login and reset on logout. Added updating of local track on track data receive.
+- Updated login/logout blocks (added extra information, profile link).
+- Added total and local played count fields in TrackInfo and in local db.
+- Added sending timestamps to the server for `postIcrementPlayedCount`, `postToggleFavorite` calls.
+- Ensured update of the current language in the settings login browser.
+- Added permanent favorite controls on track items (when not in a compact mode).
+- Published production release on Google Play.
+
+Version v.1.1.9 has been skipped (it was used for open testing release).
+
+See:
+
+- [Issue #58: Add synchronization of playback positions](https://github.com/lilliputten/march-tales-flutter-app/issues/58)
+- [Compare with previous tag](https://github.com/lilliputten/march-tales-flutter-app/compare/march-tales-app-v.0.1.8...v.0.1.10)
+- [Release](https://github.com/lilliputten/march-tales-flutter-app/releases/tag/v.0.1.10)
 
 ## [v.0.1.8](https://github.com/lilliputten/march-tales-flutter-app/releases/tag/march-tales-app-v.0.1.8) - 2025.03.19
 
@@ -25,4 +46,3 @@ The initial version is version 0.1.6 with the minimum implemented set of functio
 - Floating record player.
 - Notification player management.
 - Memorizing the playback position of each track.
-